apisample14.html

<!DOCTYPE HTML>
<title>Example with a dedicated worker</title>
<script src="testharness.js"></script>
<script src="testharnessreport.js"></script>
<script>
var worker = new Worker('apisample14-worker.js');
join_tests_on_message_port(worker);

test(function(t) {
  assert_false(false, "False is false");
}, "Test running on main document. " +
"There should be an additional test named \'Test running on worker\'.");
</script>
